+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+8.2
+---
+
+ * Add support for the `!!null`, `!!bool` and `!!int` tags, and accept quoted values for them and for `!!float`
+ * Throw a `ParseException` when the value of a `!!float` tag is not a valid float, instead of casting it to `0.0`
+
 8.0
 ---